requirement.txt
numpy
pandas
scikit-learn
librosa
praat-parselmouth
scipy
flask
joblib